+++ test/tests.cpp
@@ -48,7 +48,7 @@ TEST(ModelTests, GreedySearchGptFp32) {
   }
 
   // Verify outputs match expected outputs
-  for (int i = 0; i < params.batch_size; i++) {
+  for (size_t i = 0; i < static_cast<size_t>(params.batch_size); i++) {
     auto sequence = generator->GetSequence(i).GetCPU();
     auto* expected_output_start = &expected_output[i * params.search.max_length];
     EXPECT_TRUE(0 == std::memcmp(expected_output_start, sequence.data(), params.search.max_length * sizeof(int32_t)));
@@ -95,11 +95,11 @@ TEST(ModelTests, BeamSearchGptFp32) {
     search.SelectTop();
   }
 
-  std::vector<int32_t> output_sequence(search.params_.batch_size * search.params_.search.max_length);
+  std::vector<int32_t> output_sequence(static_cast<size_t>(search.params_.batch_size) * search.params_.search.max_length);
   search.Finalize(1, Generators::cpu_span<int32_t>{output_sequence}, {});
 
   // Verify outputs match expected outputs
-  for (int i = 0; i < search.params_.batch_size; i++) {
+  for (size_t i = 0; i < static_cast<size_t>(search.params_.batch_size); i++) {
     auto sequence = std::span<int32_t>(output_sequence.data() + search.params_.search.max_length * i, search.params_.search.max_length);
     auto* expected_output_start = &expected_output[i * search.params_.search.max_length];
     EXPECT_TRUE(0 == std::memcmp(expected_output_start, sequence.data(), params.search.max_length * sizeof(int32_t)));
